+++ src/validators/UrlValidator.php
@@ -11,9 +11,15 @@ class UrlValidator extends YiiUrlValidator
 
     public function __construct(array $config = [])
     {
+
         // Override the $pattern regex so that a TLD is not required, and the protocol may be relative and can be a #
         if (!isset($config['pattern'])) {
             $config['pattern'] = '/^(\/|#|(?:(?:{schemes}:)?\/\/(([A-Z0-9][A-Z0-9_-]*)(\.[A-Z0-9][A-Z0-9_-]*)+)?|\/)[^\s]*$)/i';
+        }
+
+        // Enable support for validating international domain names if the intl extension is available.
+        if (!isset($config['enableIDN']) && Craft::$app->getI18n()->getIsIntlLoaded() && defined('INTL_IDNA_VARIANT_UTS46')) {
+            $config['enableIDN'] = true;
         }
 
         parent::__construct($config);
@@ -21,9 +27,16 @@ class UrlValidator extends YiiUrlValidator
 
     public function validateValue($value)
     {
-        // Add support for protocol-relative URLs, # or /
-        if ($this->defaultScheme !== null && ( strpos($value, '/') === 0 || $value === '#' || $value === '/')) {
+        if($value === '#' || $value === '/') {
             $this->defaultScheme = null;
+            return null;
+        }
+
+        // Add support for protocol-relative URLs, # or /
+        if ($this->defaultScheme !== null && strpos($value, '/') === 0) {
+            $this->defaultScheme = null;
+        } else {
+            $this->pattern = '/^{schemes}:\/\/(([A-Z0-9][A-Z0-9_-]*)(\.[A-Z0-9][A-Z0-9_-]*)+)(?::\d{1,5})?(?:$|[?\/#])/i';
         }
 
         return parent::validateValue($value);